requirements.txt
# Core
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4

# Visualization
plotly==5.20.0
networkx==3.2.1
matplotlib==3.8.3

# NLP / ML
torch>=2.5.1
torchvision>=0.20.1
transformers>=5.5.4
sentence-transformers>=2.6.1
faiss-cpu==1.7.4
bertopic==0.16.0
umap-learn==0.5.6
hdbscan==0.8.33
scikit-learn==1.4.1.post1

# Community detection
python-louvain==0.16

# LLM
groq==0.5.0

# Utilities
python-dotenv==1.0.1
